+++ plugins/guests/linux/cap/mount_nfs.rb
@@ -18,8 +18,14 @@ module VagrantPlugins
             # Mount
             hostpath = opts[:hostpath].dup
             hostpath.gsub!("'", "'\\\\''")
-            mount_command = "mount -o udp,vers=#{opts[:nfs_version]} #{ip}:'#{hostpath}' #{expanded_guest_path}"
 
+            # Figure out any options
+            mount_opts = ["vers=#{opts[:nfs_version]}", "udp"]
+            if opts[:mount_options]
+              mount_opts = opts[:mount_options].dup
+            end
+
+            mount_command = "mount -o '#{mount_opts.join(",")}' #{ip}:'#{hostpath}' #{expanded_guest_path}"
             retryable(:on => Vagrant::Errors::LinuxNFSMountFailed, :tries => 5, :sleep => 2) do
               machine.communicate.sudo(mount_command,
                                        :error_class => Vagrant::Errors::LinuxNFSMountFailed)